Extract dependency rlib tests from UI harness module
Move the dependency_rlib test module out of lib_ui_tests.rs into a
separate sibling test file so the UI harness helper module stays under
the repository line-count limit.

Keep the extracted tests wired through a cfg(test) path module so they
continue to exercise the same private resolver logic without changing
coverage or behaviour.
